requirements.txt
discord.py
gTTS
PyNaCl
Flask
requests
